pyproject.toml
[tool.poetry]
name = "tengil"
version = "0.9.0"
description = "Turn Proxmox into an easy-to-manage NAS and container server through declarative configuration"
authors = ["Androidand"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/androidand/tengil"
repository = "https://github.com/androidand/tengil"
documentation = "https://github.com/androidand/tengil#readme"
keywords = ["zfs", "proxmox", "nas", "homelab", "infrastructure", "lxc", "samba", "nfs"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Filesystems",
]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
typer = "^0.9.0"
rich = "^13.0"
pydantic = "^2.0"
jinja2 = "^3.1.6"
click = ">=8.1,<8.2"
requests = "^2.31"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.0"
ruff = "^0.6.0"
pytest-cov = "^5.0"

[tool.poetry.scripts]
tengil = "tengil.cli:app"
tg = "tengil.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# --------------------------------------------------------
# STYLE AND QUALITY SETTINGS
# --------------------------------------------------------

[tool.black]
line-length = 100
target-version = ['py311']
skip-string-normalization = false

[tool.ruff]
# Match Black's formatting philosophy, handle lint and style.
line-length = 100
fix = true

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "B",   # bugbear
  "I",   # import sorting
  "UP",  # pyupgrade
  "SIM", # simplify
  "N",   # pep8-naming
]
ignore = [
  "E501"  # line length (handled by Black)
]

# Enforce per-file limits and structure discipline
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --color=yes"
testpaths = ["tests", "src/tests"]
python_files = ["test_*.py", "*_test.py"]
